+++ global.hpp
@@ -8,7 +8,7 @@
 #define SAMPLE_RATE (44100)
 #define LUT_RES (1024)
 
-inline float scale(float x, float in_min, float in_max, float out_min, float out_max)
+inline double scale(double x, double in_min, double in_max, double out_min, double out_max)
 {
     return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
